refactor: organize utils a bit better
